requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.7
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.0
Pillow==8.3.2
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
simplejson==3.17.5
sqlparse==0.4.2
uritemplate==3.0.1
urllib3==1.26.7
